Built and implemented structure prefab generating
I made a separate program to draw out and save structure prefabs with some controls over what to do with them and now I can load in structure files in this game to generate them at a position in the world. So now I need to make a few structures and scatter them around in a better way since right now it's pure random placement of diamond towers.