Menu 99%, spell/creature+, robe+, overall misc
I've cleaned some stuff up, fixed some issues with the bananas, balanced some stuff, robes now pick up and assign spells and have the sprites they should and you go back to where you died properly, not sure if that was working last time. I also made the creature class have a hit flash and other basic values so that they can take hits and respond. The knockback uses a fun little thing where it applies upwards kb every hit so that the creatures sort of knock around in the air and not just the direction you hit them.
The menu is completely done except sounds, and the credits screen which I haven't decided what to put there, and also the key inputs on the map screen are kinda wrong but still work fine, and also the controls diagram is wrong but idk the controls yet / if I'll have a real tutorial. I added a couple fonts I think look good in the menu specifically. And very specifically I made the banana mining spawns proper by checking surrounding blocks to place them in. Basically the game is pretty polished up and I'm going to add a few more enemies, a few more random logic things to the gameplay (block break effects for example) and make the generation more complex, then add in some more items and content bits plus sounds and music and reworking general balance and make the game options work, and we're mostly done. Sounds like a lot but I could have everything but the enemies knocked out in 2 days I think (finding good sounds besides), then enemies could be a day or two or a handful but eh. I'll keep it simple.