Switching macros for global, macro refs are STUPID
When you put a referenced type like an array into a macro it recreates and holds that array every time you use it, it's not like creating a fixed array to grab from at all times, it's just packing the code with duplicated arrays every time which is horrible and spikes memory use and lots of bad things. I'm doing some optimization stuff here, i'll do more in a minute but I want to see the previous builds memory use so I gotta commit here.