UI, cost (bomb count, mana), stat regen, misc
This adds ui for bomb count and mana pool and also the values for those under the hood so you have a limited pool of those resources that regenerates over time. I also made health regen over time and adjusted some values there. I made a shader for the ui but I'm not super happy with it and I added shader reloading live and also some random other finishing up things (for example flying is now disableable so when I have robes that allow or don't i can set it.) The only other thing is small adjustments to the worms and material orbs to get them to split up a bit from each other timing wise. But yeah, managerial stuff. Cool.