Sounds & polish + pause menu raw scrap to work on
This is a bunch of beautification and metal tiles and some little things and sounds but I'm also trying to get pause menus and the intro cut scene things working as intended today. That's a bit tiresome though and i'm not doing it right now, I did import the two existing menus like it to work with though.