Bunch of nonsense but also saves done now, items, robes ect
This doesn't include enemies or health status or condition stuff like that, just that items now spawn back where they were with the type they are, including robes, picks, and material nodes. Great! However I still need to fix spawn positions (better items further down) and make spells a pick up type. Which also implies adding a couple more spells, maybe a spray spell and a utility spell or something.

I also need to redo the boss, add one more flying enemy, and redo the cutscenes. Then maybe mess with the generation and we're good. That should be fine for a release, not ideal, I'll update it over time as I think of fun ideas but for a simple MVP this is good.